+++ pytorch_helpers/habana_helpers/stats.cpp
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021-2024 Intel Corporation
+ * Copyright (c) 2021-2025 Intel Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,7 @@
 #include <sstream>
 #include <string>
 #include <vector>
-#include "backend/synapse_helpers/env_flags.h"
+#include "backend/synapse_helpers/env_flags.h" // IWYU pragma: keep
 #include "logging.h"
 
 /**************************************/
@@ -40,7 +40,7 @@ StatsBase::~StatsBase() {
   // For table format, print only if at least one member is not 0
   if (m_isTbl) {
     bool allZero = true;
-    for (int i = 0; i < m_maxEnum; i++) {
+    for (size_t i = 0; i < m_maxEnum; i++) {
       uint64_t count = m_pPointData[i].count.load();
       if (count != 0) {
         allZero = false;
@@ -70,7 +70,7 @@ void StatsBase::init(
   m_pointMsg.reset(new std::string[m_maxEnum]{});
   m_pointAttributes.reset(new PointAttrMap[m_maxEnum]{});
 
-  for (int i = 0; i < m_maxEnum; i++) {
+  for (size_t i = 0; i < m_maxEnum; i++) {
     m_pointMsg[i] = names[i];
   }
 }
@@ -85,7 +85,7 @@ StatsBase::StatsBase(const StatsBase& other)
   m_pPointData.reset(new sumCollectData[m_maxEnum]{});
   m_pointMsg.reset(new std::string[m_maxEnum]{});
   m_pointAttributes.reset(new PointAttrMap[m_maxEnum]{});
-  for (int i = 0; i < m_maxEnum; i++) {
+  for (size_t i = 0; i < m_maxEnum; i++) {
     m_pointMsg[i] = other.m_pointMsg[i];
     m_pointAttributes[i] = other.m_pointAttributes[i];
   }
@@ -105,7 +105,7 @@ void StatsBase::printToLog(std::string msg, bool dumpAll, bool clear) {
   std::stringstream out;
   out << m_grep;
   std::string msgOut = "   -----  " + m_statName + " " + msg;
-  for (int i = 0; i < m_maxEnum; i++) {
+  for (size_t i = 0; i < m_maxEnum; i++) {
     uint64_t count = m_pPointData[i].count.load();
     uint64_t sum = m_pPointData[i].sum.load();
     uint64_t last_meas = m_pPointData[i].last_measurement.load();
@@ -164,7 +164,7 @@ void StatsBase::printToLog(std::string msg, bool dumpAll, bool clear) {
 void StatsBase::outputHeader() {
   std::stringstream out;
   out << m_grep;
-  for (int i = 0; i < m_maxEnum; i++) {
+  for (size_t i = 0; i < m_maxEnum; i++) {
     out << m_pointMsg[i] << "-sum,";
     out << m_pointMsg[i] << "-count,";
     out << m_pointMsg[i] << "-average,";
@@ -199,7 +199,7 @@ void StatsBase::updateEnableGlbl() {
 }
 
 void StatsBase::clearAll() {
-  for (int i = 0; i < m_maxEnum; i++) {
+  for (size_t i = 0; i < m_maxEnum; i++) {
     m_pPointData[i].count = 0;
     m_pPointData[i].sum = 0;
   }
